dev_requirements.txt
absl-py==0.9.0
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.5
asn1crypto==1.3.0
astor==0.8.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.4
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.3
django-adminlte3==0.1.6
django-crispy-forms==1.9.0
django-heroku==0.3.1
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.2.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
imagesize==1.2.0
importlib-metadata==0.19
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
numpydoc==0.9.2
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==19.1
pandas==1.0.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathlib==1.0.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
psycopg2==2.8.5
py==1.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.2
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.0.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.7.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.2
spyder-kernels==1.9.0
sqlparse==0.3.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.0
ujson==1.35
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
whitenoise==5.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
yapf==0.28.0
zipp==0.5.2
